requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Werkzeug>=3.0.1
requests>=2.32.3
librosa>=0.10.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
praat-parselmouth>=0.4.3
scipy>=1.11.0,<1.13.0
matplotlib>=3.8.0
pydub>=0.25.1
static-ffmpeg>=2.5
pytest>=8.0.0
python-dotenv>=1.0.0